refactor: 조직 내 회원 조회 member로 이동
