+++ backend/src/main/java/com/woowacourse/thankoo/alarm/application/AlarmEventListener.java
@@ -1,23 +1,34 @@
 package com.woowacourse.thankoo.alarm.application;
 
+import com.woowacourse.thankoo.alarm.application.dto.Message;
+import com.woowacourse.thankoo.alarm.domain.Alarm;
 import com.woowacourse.thankoo.common.dto.AlarmEvent;
+import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionalEventListener;
 
 @Component
 @RequiredArgsConstructor
 @Slf4j
+@Transactional
 public class AlarmEventListener {
 
     private final AlarmService alarmService;
 
-    @Async(value = "asyncTaskExecutor")
+    private final MessageFormStrategyFactory messageFormStrategyFactory;
+    private final AlarmSender alarmSender;
+
+    //    @Async(value = "asyncTaskExecutor")
     @TransactionalEventListener
     public void handle(final AlarmEvent alarmEvent) {
         log.debug("alarm event = {}", alarmEvent);
-        alarmService.send(alarmEvent.toAlarmSpecification());
+//        alarmService.send(alarmEvent.toAlarmSpecification());
+        Alarm alarm = Alarm.create(alarmEvent.toAlarmSpecification());
+        MessageFormStrategy strategy = messageFormStrategyFactory.getStrategy(alarm.getAlarmType());
+        Message message = Message.of(alarm, strategy);
+        alarmSender.send(message);
+        throw new RuntimeException();
     }
 }
