+++ backend/src/main/java/com/woowacourse/thankoo/organization/application/OrganizationQueryService.java
@@ -1,10 +1,8 @@
 package com.woowacourse.thankoo.organization.application;
 
 import com.woowacourse.thankoo.common.exception.ErrorType;
-import com.woowacourse.thankoo.organization.domain.MemberModel;
 import com.woowacourse.thankoo.organization.domain.OrganizationQueryRepository;
 import com.woowacourse.thankoo.organization.exception.InvalidOrganizationException;
-import com.woowacourse.thankoo.organization.presentation.dto.OrganizationMemberResponse;
 import com.woowacourse.thankoo.organization.presentation.dto.OrganizationResponse;
 import com.woowacourse.thankoo.organization.presentation.dto.SimpleOrganizationResponse;
 import java.util.List;
@@ -24,17 +22,6 @@ public class OrganizationQueryService {
         return organizationQueryRepository.findMemberOrganizationsByMemberId(memberId)
                 .stream()
                 .map(OrganizationResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    @Transactional(readOnly = true)
-    public List<OrganizationMemberResponse> getOrganizationMembersExcludeMe(final Long memberId,
-                                                                            final Long organizationId) {
-        List<MemberModel> members = organizationQueryRepository.findOrganizationMembersExcludeMe(
-                organizationId, memberId);
-        return members
-                .stream()
-                .map(OrganizationMemberResponse::from)
                 .collect(Collectors.toList());
     }
