+++ backend/src/test/java/com/woowacourse/thankoo/acceptance/OrganizationAcceptanceTest.java
@@ -1,12 +1,8 @@
 package com.woowacourse.thankoo.acceptance;
 
 import static com.woowacourse.thankoo.acceptance.builder.OrganizationAssured.조직_번호;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HOHO_NAME;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_NAME;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_NAME;
 import static com.woowacourse.thankoo.common.fixtures.OAuthFixture.CODE_SKRR;
-import static com.woowacourse.thankoo.common.fixtures.OAuthFixture.HOHO_TOKEN;
-import static com.woowacourse.thankoo.common.fixtures.OAuthFixture.HUNI_TOKEN;
 import static com.woowacourse.thankoo.common.fixtures.OAuthFixture.SKRR_TOKEN;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_NO_CODE;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_THANKOO_CODE;
@@ -166,38 +162,5 @@ class OrganizationAcceptanceTest extends AcceptanceTest {
                 .내_조직을_조회한다(userToken)
                 .response()
                 .조직상태가_변경됨(organizationResponse.getOrganizationId(), true);
-    }
-
-    @DisplayName("나를 제외한 조직의 모든 회원을 조회한다.")
-    @Test
-    void getMyOrganizations() {
-        TokenResponse userToken = AuthenticationAssured.request()
-                .회원가입_한다(SKRR_TOKEN, SKRR_NAME)
-                .로그인_한다(CODE_SKRR)
-                .response()
-                .body(TokenResponse.class);
-
-        TokenResponse hohoToken = AuthenticationAssured.request()
-                .회원가입_한다(HOHO_TOKEN, HOHO_NAME)
-                .token();
-
-        TokenResponse huniToken = AuthenticationAssured.request()
-                .회원가입_한다(HUNI_TOKEN, HUNI_NAME)
-                .token();
-
-        기본_조직이_생성됨();
-        OrganizationResponse organizationResponse = OrganizationAssured.request()
-                .조직에_참여한다(userToken, 조직_번호(ORGANIZATION_WOOWACOURSE_CODE))
-                .조직에_참여한다(hohoToken, 조직_번호(ORGANIZATION_WOOWACOURSE_CODE))
-                .조직에_참여한다(huniToken, 조직_번호(ORGANIZATION_WOOWACOURSE_CODE))
-                .내_조직을_조회한다(userToken)
-                .response()
-                .bodies(OrganizationResponse.class).get(0);
-
-        OrganizationAssured.request()
-                .나를_제외한_조직의_모든_회원을_조회한다(userToken, organizationResponse.getOrganizationId())
-                .response()
-                .status(HttpStatus.OK.value())
-                .나를_제외하고_모두_조회됨(2);
     }
 }
