+++ backend/src/test/java/com/woowacourse/thankoo/member/presentation/MemberControllerTest.java
@@ -5,8 +5,11 @@ import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_EMAIL;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_IMAGE_URL;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_NAME;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_SOCIAL_ID;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_EMAIL;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_NAME;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_SOCIAL_ID;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_IMAGE_URL;
+import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.createDefaultOrganization;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -22,6 +25,8 @@ import static org.springframework.restdocs.payload.JsonFieldType.STRING;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
+import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.request.RequestDocumentation.requestParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -31,13 +36,18 @@ import com.woowacourse.thankoo.member.application.dto.MemberNameRequest;
 import com.woowacourse.thankoo.member.application.dto.MemberProfileImageRequest;
 import com.woowacourse.thankoo.member.domain.Member;
 import com.woowacourse.thankoo.member.presentation.dto.MemberResponse;
+import com.woowacourse.thankoo.member.presentation.dto.OrganizationMemberResponse;
 import com.woowacourse.thankoo.member.presentation.dto.ProfileImageUrlResponse;
+import com.woowacourse.thankoo.organization.domain.Organization;
+import com.woowacourse.thankoo.organization.domain.OrganizationMember;
+import com.woowacourse.thankoo.organization.domain.OrganizationValidator;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
 
@@ -149,4 +159,48 @@ class MemberControllerTest extends ControllerTest {
                         fieldWithPath("[].imageUrl").type(STRING).description("imageUrl")
                 )));
     }
+
+    @DisplayName("본인을 제외한 모든 회원을 조회힌다.")
+    @Test
+    void getMembersExcludeMe() throws Exception {
+        given(jwtTokenProvider.getPayload(anyString())).willReturn("1");
+
+        Member huni = new Member(1L, HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID, SKRR_IMAGE_URL);
+        Member lala = new Member(2L, LALA_NAME, LALA_EMAIL, LALA_SOCIAL_ID, SKRR_IMAGE_URL);
+        OrganizationValidator organizationValidator = Mockito.mock(OrganizationValidator.class);
+        doNothing().when(organizationValidator).validate(any(Organization.class));
+        Organization organization = createDefaultOrganization(organizationValidator);
+
+        List<OrganizationMemberResponse> responses = List.of(
+                OrganizationMemberResponse.from(new OrganizationMember(huni, organization, 1, true)),
+                OrganizationMemberResponse.from(new OrganizationMember(lala, organization, 1, true))
+        );
+
+        given(memberQueryService.getOrganizationMembersExcludeMe(anyLong(), anyLong())).willReturn(responses);
+
+        ResultActions resultActions = mockMvc.perform(get("/api/members")
+                        .queryParam("organization", "1")
+                        .header(HttpHeaders.AUTHORIZATION, ACCESS_TOKEN)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andExpectAll(
+                        status().isOk(),
+                        content().string(objectMapper.writeValueAsString(responses)));
+
+        resultActions.andDo(document("organizations/get-members",
+                getResponsePreprocessor(),
+                requestHeaders(
+                        headerWithName(HttpHeaders.AUTHORIZATION).description("token")
+                ),
+                requestParameters(
+                        parameterWithName("organization").description("organizationId")
+                ),
+                responseFields(
+                        fieldWithPath("[].id").type(NUMBER).description("id"),
+                        fieldWithPath("[].name").type(STRING).description("name"),
+                        fieldWithPath("[].email").type(STRING).description("email"),
+                        fieldWithPath("[].imageUrl").type(STRING).description("imageUrl")
+                )));
+    }
+
 }
