+++ backend/src/test/java/com/woowacourse/thankoo/organization/domain/OrganizationQueryRepositoryTest.java
@@ -1,21 +1,9 @@
 package com.woowacourse.thankoo.organization.domain;
 
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HOHO_EMAIL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HOHO_IMAGE_URL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HOHO_NAME;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HOHO_SOCIAL_ID;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_EMAIL;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_IMAGE_URL;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_NAME;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_SOCIAL_ID;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_EMAIL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_IMAGE_URL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_NAME;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_SOCIAL_ID;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_EMAIL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_IMAGE_URL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_NAME;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_SOCIAL_ID;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_WOOWACOURSE_CODE;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.createDefaultOrganization;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.createThankooOrganization;
@@ -88,42 +76,6 @@ class OrganizationQueryRepositoryTest {
                 () -> assertThat(memberOrganizations.get(0).getOrderNumber()).isEqualTo(1),
                 () -> assertThat(memberOrganizations.get(1).getId()).isEqualTo(thankoo.getId()),
                 () -> assertThat(memberOrganizations.get(1).getOrderNumber()).isEqualTo(2)
-        );
-    }
-
-    @DisplayName("조직의 회원 중 내가 아닌 회원을 모두 조회한다.")
-    @Test
-    void findOrganizationByOrganizationIdNotMemberId() {
-        Member lala = memberRepository.save(
-                new Member(LALA_NAME, LALA_EMAIL, LALA_SOCIAL_ID,
-                        LALA_IMAGE_URL));
-        Member skrr = memberRepository.save(
-                new Member(SKRR_NAME, SKRR_EMAIL, SKRR_SOCIAL_ID,
-                        SKRR_IMAGE_URL));
-        Member hoho = memberRepository.save(
-                new Member(HOHO_NAME, HOHO_EMAIL, HOHO_SOCIAL_ID,
-                        HOHO_IMAGE_URL));
-        Member huni = memberRepository.save(
-                new Member(HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID,
-                        HUNI_IMAGE_URL));
-
-        Organization organization1 = organizationRepository.save(createDefaultOrganization(organizationValidator));
-        Organization organization2 = organizationRepository.save(createThankooOrganization(organizationValidator));
-
-        organization1.join(lala, new OrganizationMembers(List.of()));
-        organization1.join(skrr, new OrganizationMembers(List.of()));
-        organization1.join(hoho, new OrganizationMembers(List.of()));
-        organization2.join(huni, new OrganizationMembers(List.of()));
-
-        organizationRepository.flush();
-
-        List<MemberModel> members = organizationQueryRepository.findOrganizationMembersExcludeMe(organization1.getId(),
-                hoho.getId());
-
-        assertAll(
-                () -> assertThat(members).hasSize(2),
-                () -> assertThat(members.get(0)).extracting("id").isEqualTo(lala.getId()),
-                () -> assertThat(members.get(1)).extracting("id").isEqualTo(skrr.getId())
         );
     }
