+++ backend/src/test/java/com/woowacourse/thankoo/organization/presentation/OrganizationControllerTest.java
@@ -1,18 +1,9 @@
 package com.woowacourse.thankoo.organization.presentation;
 
-import static com.woowacourse.thankoo.common.fixtures.AuthenticationFixture.ACCESS_TOKEN;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_EMAIL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_NAME;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_SOCIAL_ID;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_EMAIL;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_NAME;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.LALA_SOCIAL_ID;
-import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_IMAGE_URL;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_THANKOO;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_THANKOO_CODE;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_WOOWACOURSE;
 import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.ORGANIZATION_WOOWACOURSE_CODE;
-import static com.woowacourse.thankoo.common.fixtures.OrganizationFixture.createDefaultOrganization;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -31,25 +22,18 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWit
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.woowacourse.thankoo.common.ControllerTest;
-import com.woowacourse.thankoo.member.domain.Member;
 import com.woowacourse.thankoo.organization.application.dto.OrganizationJoinRequest;
 import com.woowacourse.thankoo.organization.domain.MemberOrganization;
-import com.woowacourse.thankoo.organization.domain.Organization;
-import com.woowacourse.thankoo.organization.domain.OrganizationMember;
-import com.woowacourse.thankoo.organization.domain.OrganizationValidator;
 import com.woowacourse.thankoo.organization.domain.SimpleOrganization;
-import com.woowacourse.thankoo.organization.presentation.dto.OrganizationMemberResponse;
 import com.woowacourse.thankoo.organization.presentation.dto.OrganizationResponse;
 import com.woowacourse.thankoo.organization.presentation.dto.SimpleOrganizationResponse;
 import java.util.List;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
 
@@ -149,45 +133,6 @@ class OrganizationControllerTest extends ControllerTest {
                         fieldWithPath("[].lastAccessed").type(BOOLEAN).description("lastAccessed")
                 )
         ));
-    }
-
-    @DisplayName("본인을 제외한 모든 회원을 조회힌다.")
-    @Test
-    void getMembersExcludeMe() throws Exception {
-        given(jwtTokenProvider.getPayload(anyString())).willReturn("1");
-
-        Member huni = new Member(1L, HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID, SKRR_IMAGE_URL);
-        Member lala = new Member(2L, LALA_NAME, LALA_EMAIL, LALA_SOCIAL_ID, SKRR_IMAGE_URL);
-        OrganizationValidator organizationValidator = Mockito.mock(OrganizationValidator.class);
-        doNothing().when(organizationValidator).validate(any(Organization.class));
-        Organization organization = createDefaultOrganization(organizationValidator);
-
-        List<OrganizationMemberResponse> responses = List.of(
-                OrganizationMemberResponse.from(new OrganizationMember(huni, organization, 1, true)),
-                OrganizationMemberResponse.from(new OrganizationMember(lala, organization, 1, true))
-        );
-
-        given(organizationQueryService.getOrganizationMembersExcludeMe(anyLong(), anyLong())).willReturn(responses);
-
-        ResultActions resultActions = mockMvc.perform(get("/api/organizations/1/members")
-                        .header(HttpHeaders.AUTHORIZATION, ACCESS_TOKEN)
-                        .accept(MediaType.APPLICATION_JSON))
-                .andDo(print())
-                .andExpectAll(
-                        status().isOk(),
-                        content().string(objectMapper.writeValueAsString(responses)));
-
-        resultActions.andDo(document("organizations/get-members",
-                getResponsePreprocessor(),
-                requestHeaders(
-                        headerWithName(HttpHeaders.AUTHORIZATION).description("token")
-                ),
-                responseFields(
-                        fieldWithPath("[].id").type(NUMBER).description("id"),
-                        fieldWithPath("[].name").type(STRING).description("name"),
-                        fieldWithPath("[].email").type(STRING).description("email"),
-                        fieldWithPath("[].imageUrl").type(STRING).description("imageUrl")
-                )));
     }
 
     @DisplayName("조직에 접근한다.")
